tentative fix for #603: using fps in bitrate computation

+++ src/io/mp4_output.h
@@ -23,7 +23,7 @@ struct mp4_struct {
 	struct SwsContext *sws_ctx;
 	struct SwrContext *swr_ctx;
 
-	int bitrate;
+	int64_t bitrate;
 	int src_w, src_h;
 
 };
